+++ src/app/section/GallerySection/GallerySection.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Slider from "./Slider";
+import Slider from "../../components/slider/Slider";
 import styles from "./GallerySection.module.css";
 
 export default function GallerySection() {
