+++ src/components/SectionFirst.tsx
@@ -2,13 +2,12 @@
 import NavBar from "./NavBar";
 import Footer from "./Footer";
 import FeatureSection from "../app/components/FeatureSection/FeatureSection";
-import GallerySection from "./GallerySection";
+import GallerySection from "../app/section/GallerySection/GallerySection";
 import Amenities from "./Amenities";
 import NearSection from "./NearSection";
 import PaymentSection from "./PaymentSection";
 import EnquirePropertySection from "./EnquirePropertySection";
 import SimiliarProjectSection from "./SimiliarProjectSection";
-
 
 function SectionFirst() {
   return (
@@ -29,7 +28,7 @@ function SectionFirst() {
       <EnquirePropertySection />
       {/* similiar project */}
       <SimiliarProjectSection />
-       <Footer />  
+      <Footer />
     </>
   );
 }
